+++ c_step27/outer_circle.py
@@ -96,8 +96,8 @@ class OuterCircle():
 
             # 円弧
             # 楕円、描画する画像を指定、座標(x,y),xyの半径、角度,色、線の太さ(-1は塗りつぶし)
-            start_angle = math.degrees(theta) - self.unit_arc/2
-            end_angle = math.degrees(theta) + self.unit_arc/2
+            start_angle = int(math.degrees(theta) - self.unit_arc/2)
+            end_angle = int(math.degrees(theta) + self.unit_arc/2)
             if start_angle == end_angle:
                 end_angle += 1  # 差が 0 だと変なとこ描画するんで
             cv2.ellipse(canvas,
